backend/pyproject.toml
# backend/pyproject.toml
[tool.poetry]
name = "ev-charging-optimizer"
version = "0.1.0"
description = "Smart EV Charging Network Optimization for Tamil Nadu"
authors = ["Your Name <your.email@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.104.1"
uvicorn = {extras = ["standard"], version = "^0.24.0"}
pydantic = "^2.5.0"
pydantic-settings = "^2.1.0"
httpx = "^0.25.2"
aiofiles = "^23.2.1"
python-multipart = "^0.0.6"
python-dotenv = "^1.0.0"
langgraph = "^0.0.65"
langchain-core = "^0.2.0"
langchain-google-genai = "^1.0.1"
weaviate-client = "^3.25.3"
geopy = "^2.4.0"
folium = "^0.15.1"
pandas = "^2.1.4"
numpy = ">=2.1,<3.0"
requests = "^2.31.0"
overpy = "^0.7"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.3"
pytest-asyncio = "^0.23.2"
black = "^23.11.0"
isort = "^5.12.0"
flake8 = "^6.1.0"
mypy = "^1.7.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88
target-version = ['py311']

[tool.isort]
profile = "black"
line_length = 88

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true